chore: rename normalize-color to normalize-colors (umbrella discussion 480)
