+++ .circleci/verdaccio.yml
@@ -15,6 +15,11 @@ uplinks:
       maxSockets: 40
       maxFreeSockets: 10
 packages:
+  # Get @react-native/normalize-color from npm registry, since its used in deprecated-react-native-prop-types package
+  '@react-native/normalize-color':
+    access: $all
+    publish: $authenticated
+    proxy: npmjs
   # Group and isolate all local packages, avoid being proxy from outside
   '@react-native/*':
     access: $all
