+++ packages/normalize-colors/__tests__/normalizeColor-test.js
