+++ packages/normalize-colors/index.js.flow
